+++ beetsplug/mbsync.py
@@ -122,6 +122,12 @@ class MBSyncPlugin(BeetsPlugin):
 
             items = list(a.items())
 
+            if not items:
+                self._log.info(
+                    "Skipping empty album: {0}", album_formatted
+                )
+                continue
+
             # Do we have a valid MusicBrainz album ID?
             if not re.match(MBID_REGEX, a.mb_albumid):
                 self._log.info(